move stuff around to make it clearer
