+++ weave/ops_domain/artifact_type_ops.py
@@ -3,12 +3,13 @@ import typing
 from ..gql_op_plugin import wb_gql_op_plugin
 from ..api import op
 from . import wb_domain_types as wdt
-from ..partial_object import _make_alias
+
 from .wandb_domain_gql import (
     gql_prop_op,
     gql_direct_edge_op,
     gql_connection_op,
     gql_root_op,
+    _make_alias,
 )
 
 # Section 1/6: Tag Getters
