+++ weave/ops_domain/project_ops.py
@@ -9,8 +9,9 @@ from .wandb_domain_gql import (
     gql_direct_edge_op,
     gql_connection_op,
     gql_root_op,
+    make_root_op_gql_op_output_type,
 )
-from .. import partial_object
+
 from .. import weave_types as types
 from .. import errors
 from .. import input_provider
@@ -65,7 +66,7 @@ def root_all_projects_gql_resolver(gql_result):
     """,
         is_root=True,
         root_resolver=root_all_projects_gql_resolver,
-        gql_op_output_type=partial_object.make_root_op_gql_op_output_type(
+        gql_op_output_type=make_root_op_gql_op_output_type(
             "projects_500", lambda inputs: "", wdt.ProjectType
         ),
     ),
