+++ weave/ops_domain/run_ops.py
@@ -49,9 +49,9 @@ from .wandb_domain_gql import (
     gql_prop_op,
     gql_direct_edge_op,
     gql_connection_op,
+    _make_alias,
 )
 
-from ..partial_object import _make_alias
 
 from . import wb_util
 from .. import engine_trace
